code separated & few simplification
