+++ lib/github-provider.js
@@ -1,10 +1,13 @@
 'use strict';
+const moment = require('moment');
+const Color = require('color');
 
 const { handleRequest: cache } = require('./request-handler');
 const {
+  makeColorB,
   makeBadgeData: getBadgeData,
   makeLabel: getLabel,
-  getLogo
+  makeLogo: getLogo
 } = require('./badge-data');
 
 // GitHub commits since integration.
@@ -68,9 +71,57 @@ function mapGithubCommitsSince(camp, githubApiUrl, githubAuth) {
     }));
 }
 
+function mapGithubReleaseDate(camp, githubApiUrl, githubAuth) {
+  camp.route(/^\/github\/release-date\/([^\/]+)\/([^\/]+)\.(svg|png|gif|jpg|json)$/,
+  cache(function(data, match, sendBadge, request) {
+    const user = match[1];  // eg, microsoft
+    const repo = match[2];  // eg, vscode
+    const format = match[3];
+    const apiUrl = `${githubApiUrl}/repos/${user}/${repo}/releases/latest`;
+    const badgeData = getBadgeData('release date', data);
+    if (badgeData.template === 'social') {
+      badgeData.logo = getLogo('github', data);
+    }
+    githubAuth.request(request, apiUrl, {}, function(err, res, buffer) {
+      if (err != null) {
+        badgeData.text[1] = 'inaccessible';
+        sendBadge(format, badgeData);
+        return;
+      }
+      try {
+        const data = JSON.parse(buffer);
+  
+        if (!data.created_at) throw 'Project has no releases';
+  
+        const releaseDate = moment(data.created_at);
+       
+        // Pretty print release date
+        badgeData.text[1] = releaseDate.calendar(null, {
+          lastDay : '[Yesterday]',
+          sameDay : '[Today]',
+          lastWeek : '[last] dddd',
+          sameElse :  now => releaseDate.year() == now.year() ? 'D MMM' : 'D MMM YYYY'
+        });
+  
+        // Calculate color (older release is lighter/greyer color)
+        const colorB = makeColorB('#007ec6', data ); // blue
+        const baseColor = Color(colorB);
+        const weeksSinceRelease = moment().diff(releaseDate, 'weeks');
+        const factor = Math.max(Math.min(weeksSinceRelease / 25, 0.9), 0.1);
+        badgeData.colorB = baseColor.lighten(factor).desaturate(factor / 2).hex();
+        sendBadge(format, badgeData);
+      } catch(e) {
+        badgeData.text[1] = 'invalid';
+        sendBadge(format, badgeData);
+      }
+    });
+  }));
+}
+
 
 
 
 module.exports = {
-  mapGithubCommitsSince
+  mapGithubCommitsSince,
+  mapGithubReleaseDate
 };
