+++ server.js
@@ -27,8 +27,6 @@ var githubAuth = require('./lib/github-auth');
 var querystring = require('querystring');
 var prettyBytes = require('pretty-bytes');
 var xml2js = require('xml2js');
-var moment = require('moment');
-var Color = require('color');
 var serverSecrets = require('./lib/server-secrets');
 if (serverSecrets && serverSecrets.gh_client_id) {
   githubAuth.setRoutes(camp);
@@ -117,7 +115,8 @@ const {
 } = require('./lib/github-helpers');
 
 const {
-  mapGithubCommitsSince
+  mapGithubCommitsSince,
+  mapGithubReleaseDate
 } = require("./lib/github-provider");
 
 var semver = require('semver');
@@ -3326,53 +3325,7 @@ cache(function(data, match, sendBadge, request) {
 }));
 
 // GitHub release date integration.
-camp.route(/^\/github\/release-date\/([^\/]+)\/([^\/]+)\.(svg|png|gif|jpg|json)$/,
-cache(function(data, match, sendBadge, request) {
-  var user = match[1];  // eg, qubyte/rubidium
-  var repo = match[2];
-  var format = match[3];
-  var apiUrl = githubApiUrl + '/repos/' + user + '/' + repo + '/releases/latest';
-  var badgeData = getBadgeData('release date', data);
-  if (badgeData.template === 'social') {
-    badgeData.logo = getLogo('github', data);
-  }
-  githubAuth.request(request, apiUrl, {}, function(err, res, buffer) {
-    if (err != null) {
-      badgeData.text[1] = 'inaccessible';
-      sendBadge(format, badgeData);
-      return;
-    }
-    try {
-      var data = JSON.parse(buffer);
-
-      if (typeof data.created_at !== 'string') throw 'Project has no releases';
-
-      // Pretty print release date
-      var releaseDate = moment(data.created_at);
-      var dateString = releaseDate.calendar(null, {
-        lastDay : '[Yesterday]',
-        sameDay : '[Today]',
-        lastWeek : '[last] dddd',
-        sameElse: 'D MMM YYYY'
-      });
-      // Trim current year from date string
-      var currentYear = moment().year();
-      badgeData.text[1] = dateString.replace(' ' + currentYear, '');
-
-      // Calculate color (older release is lighter/greyer color)
-      const colorB = '#007ec6'; // blue
-      const baseColor = Color(colorB);
-      const weeksSinceRelease = moment().diff(releaseDate, 'weeks');
-      const factor = Math.max(Math.min(weeksSinceRelease / 25, 0.9), 0.1);
-      badgeData.colorB = baseColor.lighten(factor).desaturate(factor / 2).hex();
-
-      sendBadge(format, badgeData);
-    } catch(e) {
-      badgeData.text[1] = 'none';
-      sendBadge(format, badgeData);
-    }
-  });
-}));
+mapGithubReleaseDate(camp, githubApiUrl, githubAuth)
 
 // GitHub commits since integration.
 mapGithubCommitsSince(camp, githubApiUrl ,githubAuth);
